requirements.txt
google-api-python-client
pynput
python-dotenv
